+++ src/Facades/Kafka.php
@@ -7,12 +7,16 @@ use Junges\Kafka\Message\Message;
 use Junges\Kafka\Support\Testing\Fakes\KafkaFake;
 
 /**
- * @method static \Junges\Kafka\Contracts\CanProduceMessages publishOn(string $broker, string $topic);
- * @method static \Junges\Kafka\Consumers\ConsumerBuilder createConsumer(string $brokers, array $topics = [], string $groupId = null);
- * @method static void assertPublished(Message $message);
- * @method static void assertPublishedOn(string $topic, Message $message, $callback = null)
+ * @method static \Junges\Kafka\Contracts\CanProduceMessages publishOn(string $topic, string $broker = null);
+ * @method static \Junges\Kafka\Consumers\ConsumerBuilder createConsumer(array $topics = [], string $groupId = null, string $brokers = null);
+ * @method static void assertPublished(Message $message = null);
+ * @method static void assertPublishedTimes(int $times = 1, Message $message = null);
+ * @method static void assertPublishedOn(string $topic, Message $message = null, $callback = null)
+ * @method static void assertPublishedOnTimes(string $topic, int $times = 1, Message $message = null, $callback = null)
  * @method static void assertNothingPublished()
  * @mixin \Junges\Kafka\Kafka
+ *
+ * @see \Junges\Kafka\Kafka
  */
 class Kafka extends Facade
 {