+++ tests/KafkaFakeTest.php
@@ -28,7 +28,7 @@ class KafkaFakeTest extends LaravelKafkaTestCase
 
     public function testItStorePublishedMessagesOnArray()
     {
-        $producer = $this->fake->publishOn('broker', 'test_topic_1')
+        $producer = $this->fake->publishOn('topic')
             ->withBodyKey('test', ['test'])
             ->withHeaders(['custom' => 'header'])
             ->withKafkaKey(Str::uuid()->toString());
@@ -46,7 +46,7 @@ class KafkaFakeTest extends LaravelKafkaTestCase
             $this->assertThat($exception, new ExceptionMessage('The expected message was not published.'));
         }
 
-        $producer = $this->fake->publishOn('broker', 'topic')
+        $producer = $this->fake->publishOn('topic')
             ->withBodyKey('test', ['test'])
             ->withHeaders(['custom' => 'header'])
             ->withKafkaKey(Str::uuid()->toString());
@@ -55,9 +55,32 @@ class KafkaFakeTest extends LaravelKafkaTestCase
         $this->fake->assertPublished($producer->getMessage());
     }
 
+    public function testAssertPublishedTimes()
+    {
+        try {
+            $this->fake->assertPublished(new Message('foo'));
+        } catch (ExpectationFailedException $exception) {
+            $this->assertThat($exception, new ExceptionMessage('The expected message was not published.'));
+        }
+
+        $producer = $this->fake->publishOn('topic')
+            ->withBodyKey('test', ['test'])
+            ->withHeaders(['custom' => 'header'])
+            ->withKafkaKey(Str::uuid()->toString());
+        $producer->send();
+
+        $this->fake->assertPublishedTimes(1, $producer->getMessage());
+
+        try {
+            $this->fake->assertPublishedTimes(2, new Message('foo'));
+        } catch (ExpectationFailedException $exception) {
+            $this->assertThat($exception, new ExceptionMessage('Kafka published 1 messages instead of 2.'));
+        }
+    }
+
     public function testItCanPerformAssertionsOnPublishedMessages()
     {
-        $producer = $this->fake->publishOn('broker', 'topic')
+        $producer = $this->fake->publishOn('topic')
             ->withBodyKey('test', ['test'])
             ->withHeaders(['custom' => 'header'])
             ->withKafkaKey($uuid = Str::uuid()->toString());
@@ -86,7 +109,7 @@ class KafkaFakeTest extends LaravelKafkaTestCase
 
     public function testAssertPublishedOn()
     {
-        $producer = $this->fake->publishOn('broker', 'topic')
+        $producer = $this->fake->publishOn('topic')
             ->withBodyKey('test', ['test'])
             ->withHeaders(['custom' => 'header'])
             ->withKafkaKey(Str::uuid()->toString());
@@ -104,9 +127,29 @@ class KafkaFakeTest extends LaravelKafkaTestCase
         }
     }
 
+    public function testAssertPublishedOnTimes()
+    {
+        $producer = $this->fake->publishOn('topic')
+            ->withBodyKey('test', ['test'])
+            ->withHeaders(['custom' => 'header'])
+            ->withKafkaKey(Str::uuid()->toString());
+
+        $producer->send();
+
+        $this->fake->assertPublished($producer->getMessage());
+
+        $this->fake->assertPublishedOnTimes('topic', 1, $producer->getMessage());
+
+        try {
+            $this->fake->assertPublishedOnTimes('topic', 4, $producer->getMessage());
+        } catch (ExpectationFailedException $exception) {
+            $this->assertThat($exception, new ExceptionMessage('Kafka published 1 messages instead of 4.'));
+        }
+    }
+
     public function testICanPerformAssertionsUsingAssertPublishedOn()
     {
-        $producer = $this->fake->publishOn('broker', 'topic')
+        $producer = $this->fake->publishOn('topic')
             ->withBodyKey('test', ['test'])
             ->withHeaders(['custom' => 'header'])
             ->withKafkaKey($uuid = Str::uuid()->toString());